chambre_service/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary 